% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{praiserror-package}
\alias{praiserror}
\alias{praiserror-package}
\title{praiserror}
\description{
A Sarcastic Error Handler
}
\details{
This package provides you with a useless and demoralizing error handler. Please turn your speakers on.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}

